requirements.txt
mysqlclient~=2.1.0
mysql~=0.0.3
mysql-connector-python~=8.0.27
Flask~=2.0.2
Flask-MySQLDb~=0.2.0
Flask-Cors~=3.0.10
Pillow~=8.4.0
APScheduler~=3.8.1
python-dotenv~=0.19.2